DECamera: Change type for size and timeout.
Use size_t, rather than long, for buffer sizes.
Use unsigned long, rather than size_t, for time (seconds).

In addition to being the "correct way", this eliminates compiler warnings
due to type mismatches.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10915 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/DirectElectron/src/DEClientLib/DEProtoProxy.cpp
@@ -32,13 +32,13 @@ bool DEProtoProxy::close()
 	return this->server->close();
 }
 
-void DEProtoProxy::set_ParamTimeout(size_t seconds) 
+void DEProtoProxy::set_ParamTimeout(unsigned long seconds)
 { 
 	boost::lock_guard<boost::mutex>(*this->server->getTransactionMutex());
 	this->paramTimeout = seconds; 
 }
 
-void DEProtoProxy::set_ImageTimeout(size_t seconds) 
+void DEProtoProxy::set_ImageTimeout(unsigned long seconds)
 { 
 	boost::lock_guard<boost::mutex>(*this->server->getTransactionMutex());
 	// Always wait a minimum paramTimeout (or 5 secs) when setting the
